react-shortcuts: export Provider for enabling external ShortcutProvider implementors
The default `ShortcutProvider` is great for most React applications, but when
React is being rendered into multiple locations on a page (for example, in a
browser extension that extends multiple different locations on a page with
React), the default `ShortcutProvider` does not work as it is intended to be a
singleton.

In order to implement one's own `ShortcutProvider` for this use case, the `ContextProvider`
must be exported.