Use specified timezone, not local timezone, for humanizeDate()'s weekday
```
Wednesday September 4th at 23:30 in Pago Pago (American Samoa),
Thursday September 5th at 10:30 in Accra (Ghana), and
Friday September 6th at 00:30 in Kiritimati (Kiribati)
```
are all exactly the same moment in time.

Thus, a single timestamp can have a different day-of-the-week depending
on the timezone in which it is observed.

JavaScript's Date.getDay() returns the day of the week in the
"local" (usually browser) timezone.

I18n.formatDate() was calling I18n.humanizeDate(), which looked up the
day of the week using date.getDay().

But... I18n.formatDate() supports the Intl.DateTimeFormat options,
which include `timeZone:`.

If a `timeZone:` was specified for `formatDate()`, and that timezone
and the "local" timezone disagreed about which day of the week that
date was, then the result returned by `formatDate()` would be wrong.

This became apparent when a unit test passed on my local machine, but
failed in CI.

This change enables that unit test to pass regardless of what the "local"
timezone is when it is being run.