:sparkles: Transparently support external urls with Link
When a URL is rendered with the Link component, render it as a plain
`<a>` tag rather than a `<ReactRouterLink>` if the URL is external (ie.
not relative).

This ensures that we can continue to use the `<Link>` component without
needing to provide workarounds in each service to first check the URL
and resolve it differently.