Update exclude of tsbuildinfo files
The lerna v3 to v4 update included a major update to `npm-packlist`,
which controls what files get packaged into the tar file when
publishing. In npm-packlist v2 `*.tsbuildconfig` no longer excludes the
`build/tsconfig.tsbuildconfig` file, we need to give the fuller path to
ensure this file is not uploaded to the npm registry.