Update loom to v1 (#2150)
* Update loom to v1

Babel gets updated along the way. Build output is borderline identical -
there's a small change to some babel helper functions, but code remains
functionally identical

* Update node CI version to 12.22

We still claim support for 12.14 and up, but loom v1 requires 12.22 so
thats the min version we can test against.

We'll be dropping support for node 12 soon I'm not to worried by this.