Improvements to the way we await the act scope
Ensure that we're awaiting the promise within the scope of the act block
and use the flushSync API to mimic the behaviour of React 17 to
synchronously flush state to the render rather than trying to batch it.
Without the use of flushSync, we were seeing issues in shopify/web tests
where a previously tested component that left a promise in the act queue
could cause subsequent mounts in other tests to fail.