Fix subtle bug in react-form-state and react-form that requires flushSync
FormState's validateForm method makes use of the second argument to
setState which gets called back when the state is set, we await this
value as part of the submission process before proceeding so that we
know the state will be updated and flushed to render. This conflicts
with our tests where we want to await the act that wraps submit because
the act will never be able to complete because it will get stuck on the
validatForm call which never calls back because it is trying to batch
the state updates.

We wrap these calls in flushSync in order to make it work correctly.