Ignore negative start values for lifecycle events
Some browsers such as Firefox have the potential to send negative
time-to-first-byte values because of a bug in the browser:
https://bugzilla.mozilla.org/show_bug.cgi?id=1429422#c12

This is causing the aggregations that we perform on our performance data
to be skewed because we're getting nonsensical values, it doesn't make
sense for us to be recording negative values for lifecycle events
because they can't go backwards in time.

This change will drop those values, rather than report them. We'd rather
drop them than clamp them to zero because zero also isn't a relevant
indicator here.