Remove mounted ref from graphql-document
1. React no longer recommends this pattern and has removed the warning
about "set state after unmounted may indicate a memory leak" because
the intention was to help developers find places to cancel
asynchronous actions but many can't be cancelled so this just leads
to noise in the code.

2. This led to race conditions where a load() call is made while mounted
is still false and the loaded result is never set in the value. The
callback is never invoked a second time, so the value never gets set
despite the load completing.