Account for quirk of React Router v6 in navigations
We mark a Navigation as complete when either the replaceState or
pushState history APIs are invoked. This presents an issue for React
Router v6 which invokes replaceState on initialization to setup a piece
of state that it can use to track changes.

This workaround prevents us from "ending" a naviation prematurely if
replaceState was invoked only to change the "idx" state value from
null/undefined to zero which is the unique behaviour that React Router
adds.

In the future, we'll need to evolve this further and likely provide
hooks for @shopify/performance to be customized with an understanding of
alternative navigation approaches such as fetch-then-render.