Only report finished navigations by default
We send all navigations in the report currently, including cancelled and
timed out metrics. @shopify/koa-performance then receives these and
sends a navigation_complete metric for each of them which can cause the
metrics to be skewed because they also include cancelled and timed out
metrics.

This feels like a foot gun, and a likely source of errors for other
applications as well, so we're introducing a new default to only report
on navigations that are Finished, and to require the user to explicitly
override this if they also want to collect cancelled and timed out
navigations.