StatsD: consume STATSD_DEFAULT_TAGS env var
At Shopify we use the STATSD_DEFAULT_TAGS and consume this env var in most of our libraries, so far the quilt library has not consumed this environment variable. In this PR we introduce a change in the statsd client which will try to read globalTags from the environment variable, but will not override the existing values in case the user passed already tags in the constructor.

Signed-off-by: Pedro Tanaka <pedro.tanaka@shopify.com>