Update and Map Deprecated Timezone (#2842)
Correctly support deprecated timezones in older browsers when calling `formatDate()`

For example, modern browsers support both `Europe/Kyiv` and (the now deprecated) `Europe/Kiev`, but browsers as recent as Chrome 131 on MacOS only support `Europe/Kiev`.

Note: This is a purely internal change which should not effect the result of calling `formatDate()`.

Co-authored-by: Trish Rempel <trish.rempel@shopify.com>
Co-authored-by: Christina Tran <christina.tran@shopify.com>
Co-authored-by: Jess Telford <jess+github@jes.st>